Split query plane storage impls
